requirements.txt
gphoto2==2.5.0
python-dateutil
pydub 